pi_zero/requirements.txt
picamera2
RPi.GPIO
flask
opencv-python
numpy
